Tie mounting holes to VCC
The outer case of the battery is on the positive battery terminal, so
instead of having the screw holes (which sit right underneath the
battery holder) at ground, tie them to VCC to avoid the chance of a
short.

This also fills areas around the holes with VCC, leaving the center of
the board with a ground fill.